autofill: disable autofill entry and autofill download manager
With Autofill, I argue, users are tempted to be lazy and fill forms
with more data than is perhaps really required for a particular
service.

The thought here is that the user should be readily subconciously
reminded about it through every keystroke of information he is about
to enter.

Disables the autofill download manager (trk:158).