prefs: disable "Use a prediction service to help complete searches..."
This autocomplete service would send search queries as soon as you
started typing, even if not yet done typing. Some regard this as an
information leak, others as a waste of bytes sent over the network.
Either reason is fine for me to disable this by default.

Impl. details:
Set suggest_enabled to false by default.