search_engines: add DuckDuckGo as selectable and default search engine
Continue the “As long as it isn't Google…” series :-)

There is a lot of code pertaining to DefaultSearchEngine,
default_search_engine, profile resetting, etc. etc.

The default default's seem to default to index 0 in the
"engines_default" (or language-specific engines_XX) array, so
prepending &duckduckgo yields the desired result, for now at least.