src: flag evil connections
Replace URLs to Google services by URLs to our own server, so as to
analyze where we still have to patch the browser to make it stop
blurting data out. A unique hostname is used for easier identification
in our webserver logs and source code. This also facilitates watching
with tcpdump/wireshark a bit, as the DNS queries will stand out.