src: flag Google connections
Change the hostnames of Google services to contain some unique
identifier with which we can more easily distinguish HTTP CONNECT
requests appearing in Squid logs and bare DNS requests in a pcap
stream from one another.

As a side effect, most Google services are now defunct and not
usable, even if activated in preferences.