src: flag evil connections
Change the hostnames of Google services to contain some unique
identifier (xx.trk) which makes requests easier to spot in a local
Squid access.log and pcap dumps with bare DNS requests.

Also add our domain into it, to feed it through a proxy for
application-level analysis of HTTPS connections.