src: flag evil connections
Prefix URLs to Google services with trk: so that whenever something
tries to load them, a modal dialog will inform about this infraction.

If you see such dialog, we know that (a) either the URL needs to be
whitelisted, or (b) the feature that triggered it needs to be disabled
by default.