src: add comments for evil-looking code
A trk:193 alert would show whenever the Omnibar is set to use Google
as search engine. When it does, it tries to find the most suitable TLD
for your request.

Well, since we do not use Google as default search engine, this entry
can be whitelisted, i.e. have no trk: prefix.