iridium: help identify possibly evil hardcoded URLs
The Chromium codebase has left us with a number of suspect URLs, and
we want to know if the browser attempts to contact those sites.

This patch introduces a new scheme, "trk:", which, when attempted to
being processed, will dump a warning onto the screen as the resource
is loaded. All URLs we think are suspect are "blacklisted" by
prepending the new scheme to an existing URL:

	trk:1234:https://clients4.google.com/ (or)
	trk:https://clients4.google.com/

At seeing a warning, we then know that investigation needs to be done,
either to (a) whitelist the URL, that is, remove the trk: prefix and
not show the warning, and/or (b) disable the particular feature which
caused the loading of the URL in the first place, by default.

Implementation:

We hack up the URLFetcher class which sits in the network stack, and
most of the URL that get loaded pass through here. The trk: prefix is
stripped and processing continues with the inner URL. There is also a
browser-level scheme handler for trk:, not sure if we still need it
(relict from 43.0~test2). It might trigger.