net: add "trk:" scheme and help identify URLs being retrieved
The Chromium codebase has left us with a number of suspect URLs, and
we want to know if the browser attempts to contact those sites.

This patch introduces a new scheme, "trk:", which, when attempted to
being processed, will dump a warning onto the screen as the resource
is loaded. All URLs we think are suspect are "blacklisted" by
prepending the new scheme to an existing URL:

	trk:1234:https://clients4.google.com/
	trk:https://clients4.google.com/ (unnumbered old variant)
	trk:0.1234:https://... (stderr only, no UI reporting)

Upon seeing a warning, we then know to investigate further, and either
(a) whitelist the URL, that is, remove the trk: prefix and not show
the warning, and/or (b) disable the particular feature which caused
the loading of the URL in the first place, by default.

Implementation:

We hack up the URLFetcher class which sits in the network stack, and
most of the URL that get loaded pass through here. The trk: prefix is
stripped and processing continues with the inner URL.