iridium: help identify possibly evil hardcoded URLs
The Chromium codebase has left us with a number of suspect URLs, and
we want to know if the browser attempts to contact those sites.

This patch introduces a new scheme, "trk:", which, when attempted to
being processed, will dump a warning onto the screen as the resource
is loaded. All URLs we think are suspect are "blacklisted" by
prepending the new scheme to an existing URL:

	trk:https://iridiumbrowser.de/

At seeing a warning, we then know that investigation needs to be done,
either to (a) whitelist the URL, that is, remove the trk: prefix and
not show the warning, and/or (b) disable the particular feature which
caused the loading of the URL in the first place, by default.

Implementation:

Tried a few; we are using a hook in the URL fixer now.
The scheme handler is unused, but preserved for one release as
historic reference.