iridium: graphic output on trk: URL
This commit builds upon the previous one by enhancing the printf
statement (which is seldomly visible) by a graphical note displayed on
all browser windows (because we do not know which window issues the
URL request, if any).

Implementation:

The Javascript interpreter is called to execute an alert() call, which
seems both dead simple, and is a lot more reliable than trying to
spawn and manage a BubbleDelegateView class.