common: set Iridium version and change User-Agent accordingly
Reduced version number:
We do not use @BUILD@ or @PATCH@, in particular not in the User-Agent
string.

As for the user agent, continue providing Chrome/* for possible
compatibility checks by the browser and/or websites, since we
really are still Chromium.

Note to self:
Update "Chromium/*" in UA string when updating chrome/version.